Added the generalized-alpha integration method.
Added support for the generalized-alpha integration method.  The
responsibilities of some classes were changed to meet the
requirements of the new method.

The `integrator::operator()` method now accepts both the equations
of motion and the structure as arguments.  This makes it possible
for an integrator to request forces applied to the structure.

A `compute_forces` method was added to the `eom` class.  This can
be used by an integrator to calculate the force vector from the
current generalized displacement, velocity, and acceleration
vectors.

The `runner` class no longer computes the forces at the current
t(n+1) time step.  This is now the responsibility of the integrator.