Added documentation.
Added documentation to the project.  The documentation uses Pandoc
and LaTeX to generate HTML and PDF documentation.

Minor changes were made to the `about` class to enhance the
legibility of the strings that it generates.

The usage statement in `clp` was changed: `input-file` replaces
`input` as an argument.

The damping matrix in the equations of motion now defaults to a zero
matrix instead of an identity matrix.

The XML reader `input_reader` now only sets the generalized mass,
damping, and stiffness matrices if they are present in the input
file.  Otherwise, they are left at their default values.

The `cout_q` output filter was refactored as the `summary` filter.