Added a `no_header` option to the modes output filter.
Added a `no_header` option to the modes output filter.  This makes
it easier to import the data file into software other than Tecplot.

Replaced calls to `std::real()` with calls to `::yamss::real()`.
This was missed when the `::yamss::real()` method was first added
to remove dependency on C++11 compiler support.