Added an ability to run in server mode.
The YAMSS executable will now run as a server if the `--server`
option is present on the command line.  The program can still be
run in stand-alone mode, as before.  The server mode uses the
Zenith library and can communicate with Zenith clients.

The server mode is a work in progress.  Amongst other known issues,
the Thrift IDL file is not yet being installed and the server does
not yet support complex mode.

All of the inspectors were modified so that they create their
files relative to a given parent directory.  The inspectors can
also now be queried to return a set of all of the files that it
has created.

The main program has also been decomposed for better readibility.