Revised loads so that they are now applied to elements.
Loads are now applied to a set of elements instead of a set of
nodes.  Functionally, the forces and moments are still evaluated
at nodes, with those nodes being the sest of element vertices.

A point element type, removed in an earlier version, was re-added
so that point loads could be supported.

The documentation and examples were updated to reflect the changes
in the way loads are defined.