Changed handling of streams in the summary and modes inspectors.
Deleted the `stream_inspector<T>` class.  An `ostream` class was
added to the `yamss::inspector` namespace to handle situations in
which information should be sent to a file if a filename is
provided, or to the standard output console if a filename is not
provided.  The `summary` and `modes` inspectors were updated to
use this class.