Added the ability to automatically generate quadrilateral grids.
Added support for a `<grid>` element in the `<structure>` section
of the input file.  This can be used to easily create a surface
grid of quadrilateral elements.

Also updated the `motion` inspector so that it can create output
files in the Stanford Polygon File Format (.ply), as well as in
the previously supported Tecplot file format.