Added ability to define a mode shape by using analytic functions.
Modified the input file parser to allow a mode shape to be defined
by a set of analytic functions.  The functions are given as strings
and evaluated by using a Lua kernel.  The manual was updated to
describe the use of this method.

Minor changes where made to other classes to correct errors or to
enhance capability.

An example was added for modeling a square plate that is simply
supported on all sides.