Add SWIG wrapper to the shared library target
The C++ wrapper source file generated by SWIG was not being included in the shared library target.  This commit makes the necessary changes to the CMake instructions to include the wrapper code.

Also, the wrapper handler would fail when trying to create a temporary directory to store job-related files because the parent directory was missing.  The source code was changed to create that parent directory.