Code changes and bug fixes for Java integration
The SWIG interface definition file and the wrapper source were changed to fix problems with Java integration.

The `finalize()` function in the handler now checks the job URL to see if it’s protocol is `file`.  If so, it will copy output files from the job directory to the base path of the URL.  Otherwise it will do nothing.